requirements.txt
flask==3.0.3
flask-cors==4.0.1
werkzeug==3.0.3
pymongo==4.10.1
pyjwt==2.9.0
python-dotenv==1.0.1
